start with processing the singing output for legacy struct
